HAL
Squuid (incomplete) is a hardware abstraction layer for I/O (sensors and actuators) and communication (UART, Parallel, 1-wire). My aim for squuid is to make it a one-stop shop for my embedded systems projects by supporting a wide range of sensors and actuators and providing peripheral communication. The definitions for each microcontroller, register addresses, and register mask values will be saved in an EEPROM using Squuid’s memory map.